+++ src/report/chart/tooltip/test-span/badges.tsx
@@ -3,6 +3,7 @@
  */
 import { ChartSpan } from '../../../data/tests.js';
 import { formatTime } from '../../../utils.js';
+import { useReportData } from '../../state/report-data.js';
 import { Badge } from '../badge.js';
 
 type Field = {
@@ -11,12 +12,11 @@ type Field = {
 };
 
 export function Badges({ chartSpan }: { chartSpan: ChartSpan }) {
+  const { runInfo } = useReportData();
   const fields: Field[] = [
     ...getDurationBadge(chartSpan),
     ...getRetryBadge(chartSpan),
-    // for debug
-    { label: 'Worker Index', value: chartSpan.test.workerIndex },
-    { label: 'Parallel Index', value: chartSpan.test.parallelIndex },
+    ...(runInfo.debug ? getDebugBadges(chartSpan) : []),
   ];
 
   const rows: Field[][] = [];
@@ -51,4 +51,11 @@ function getDurationBadge(chartSpan: ChartSpan): Field[] {
 function getRetryBadge(chartSpan: ChartSpan): Field[] {
   if (chartSpan.test.retry === 0) return [];
   return [{ label: 'Retry', value: chartSpan.test.retry }];
+}
+
+function getDebugBadges(chartSpan: ChartSpan) {
+  return [
+    { label: 'Worker Index', value: chartSpan.test.workerIndex },
+    { label: 'Parallel Index', value: chartSpan.test.parallelIndex },
+  ];
 }